Response to Ember Object Diagnostic
Comfort: 0
Clarity: 2

I used Ember-Cli, ember-object repo, ember-route-dynamic repo
and I can't find any examples of creating classes or objects
or in what directories I should put the code anyways. I only
know how to make a route and a component, and the files only
come with a route.js and a template.hbs and objects and classes
don't seem to go in either of those files so I am lost.